liveblogs: move header to separate component
